refactor: remove config discovery compatibility layers
