+++ cmd/rslint/ignore_conformance_test.go
@@ -10,6 +10,7 @@ import (
 	"github.com/microsoft/typescript-go/shim/tspath"
 	api "github.com/web-infra-dev/rslint/internal/api"
 	rslintconfig "github.com/web-infra-dev/rslint/internal/config"
+	"github.com/web-infra-dev/rslint/internal/config/discovery"
 )
 
 func TestCLIAndAPIIgnoreConformance(t *testing.T) {
@@ -218,46 +219,23 @@ func TestCLIMultiConfigGitignoreIsolation(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	entry := map[string]any{
-		"files": []string{"**/*.ts"},
-		"rules": map[string]string{"no-debugger": "error"},
-	}
-	payload, err := json.Marshal(map[string]any{
-		"configs": []any{
-			map[string]any{
-				"configDirectory": firstDir,
-				"entries":         []any{entry},
-				"targetFiles":     []string{firstTarget},
-				"explicitOnly":    true,
-			},
-			map[string]any{
-				"configDirectory": secondDir,
-				"entries":         []any{entry},
-				"targetFiles":     []string{secondTarget},
-				"explicitOnly":    true,
-			},
+	entry := rslintconfig.RslintConfig{{
+		Files: []string{"**/*.ts"},
+		Rules: rslintconfig.Rules{"no-debugger": "error"},
+	}}
+	catalog := &discovery.ConfigCatalog{
+		Configs: map[string]rslintconfig.RslintConfig{
+			tspath.NormalizePath(firstDir):  entry,
+			tspath.NormalizePath(secondDir): entry,
 		},
-	})
-	if err != nil {
-		t.Fatal(err)
+		Scopes: map[string]rslintconfig.LintDiscoveryScope{
+			tspath.NormalizePath(firstDir):  {Files: []string{tspath.NormalizePath(firstTarget)}, ExplicitOnly: true},
+			tspath.NormalizePath(secondDir): {Files: []string{tspath.NormalizePath(secondTarget)}, ExplicitOnly: true},
+		},
 	}
-	payloadPath := filepath.Join(workspace, "config-payload.json")
-	if err := os.WriteFile(payloadPath, payload, 0o644); err != nil {
-		t.Fatal(err)
-	}
-	stdin, err := os.Open(payloadPath)
-	if err != nil {
-		t.Fatal(err)
-	}
-	originalStdin := os.Stdin
-	os.Stdin = stdin
-	t.Cleanup(func() {
-		os.Stdin = originalStdin
-		_ = stdin.Close()
-	})
 
 	code, stdout, stderr := runLintPipelineForTest(t, workspace, lintArgs{
-		ConfigStdin:    true,
+		ConfigCatalog:  catalog,
 		Format:         "jsonline",
 		NoColor:        true,
 		SingleThreaded: true,
@@ -291,24 +269,22 @@ func TestCLIExplicitOnlyConfigDoesNotBlockParentGitignore(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	payload := &parsedPayload{
-		ConfigMap: map[string]rslintconfig.RslintConfig{
-			workspace:  {{Rules: rslintconfig.Rules{"no-console": "error"}}},
-			ignoredDir: {{Rules: rslintconfig.Rules{"no-debugger": "error"}}},
+	catalog := &discovery.ConfigCatalog{
+		Configs: map[string]rslintconfig.RslintConfig{
+			tspath.NormalizePath(workspace):  {{Rules: rslintconfig.Rules{"no-console": "error"}}},
+			tspath.NormalizePath(ignoredDir): {{Rules: rslintconfig.Rules{"no-debugger": "error"}}},
 		},
-		ConfigTargetScopes: map[string]rslintconfig.LintDiscoveryScope{
-			ignoredDir: {
-				Files:        []string{explicitTarget},
+		Scopes: map[string]rslintconfig.LintDiscoveryScope{
+			tspath.NormalizePath(ignoredDir): {
+				Files:        []string{tspath.NormalizePath(explicitTarget)},
 				ExplicitOnly: true,
 			},
 		},
-		IsMultiConfig: true,
 	}
 	code, stdout, stderr := runLintPipelineForTest(t, workspace, lintArgs{
-		ConfigStdin:    true,
-		ConfigPayload:  payload,
-		AllowFiles:     []string{explicitTarget},
-		AllowDirs:      []string{ignoredDir},
+		ConfigCatalog:  catalog,
+		AllowFiles:     []string{tspath.NormalizePath(explicitTarget)},
+		AllowDirs:      []string{tspath.NormalizePath(ignoredDir)},
 		Format:         "jsonline",
 		NoColor:        true,
 		SingleThreaded: true,
@@ -345,32 +321,28 @@ func TestCLIMultiConfigGitignoreOwnershipBoundaries(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	entry := map[string]any{
-		"files": []string{"**/*.ts"},
-		"rules": map[string]string{"no-debugger": "error"},
-	}
-	payload, err := json.Marshal(map[string]any{
-		"configs": []any{
-			map[string]any{
-				"configDirectory": workspace,
-				"entries":         []any{entry},
-				"targetFiles":     []string{parentOwnedTarget, parentIgnoredTarget},
-				"explicitOnly":    true,
+	entry := rslintconfig.RslintConfig{{
+		Files: []string{"**/*.ts"},
+		Rules: rslintconfig.Rules{"no-debugger": "error"},
+	}}
+	catalog := &discovery.ConfigCatalog{
+		Configs: map[string]rslintconfig.RslintConfig{
+			tspath.NormalizePath(workspace): entry,
+			tspath.NormalizePath(childDir):  entry,
+		},
+		Scopes: map[string]rslintconfig.LintDiscoveryScope{
+			tspath.NormalizePath(workspace): {
+				Files: []string{
+					tspath.NormalizePath(parentOwnedTarget),
+					tspath.NormalizePath(parentIgnoredTarget),
+				},
+				ExplicitOnly: true,
 			},
-			map[string]any{
-				"configDirectory": childDir,
-				"entries":         []any{entry},
-				"targetFiles":     []string{childOwnedTarget},
-				"explicitOnly":    true,
+			tspath.NormalizePath(childDir): {
+				Files:        []string{tspath.NormalizePath(childOwnedTarget)},
+				ExplicitOnly: true,
 			},
 		},
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-	payloadPath := filepath.Join(workspace, "config-payload.json")
-	if err := os.WriteFile(payloadPath, payload, 0o644); err != nil {
-		t.Fatal(err)
 	}
 	for _, test := range []struct {
 		name           string
@@ -380,19 +352,8 @@ func TestCLIMultiConfigGitignoreOwnershipBoundaries(t *testing.T) {
 		{name: "concurrent"},
 	} {
 		t.Run(test.name, func(t *testing.T) {
-			stdin, err := os.Open(payloadPath)
-			if err != nil {
-				t.Fatal(err)
-			}
-			originalStdin := os.Stdin
-			os.Stdin = stdin
-			t.Cleanup(func() {
-				os.Stdin = originalStdin
-				_ = stdin.Close()
-			})
-
 			code, stdout, stderr := runLintPipelineForTest(t, workspace, lintArgs{
-				ConfigStdin:    true,
+				ConfigCatalog:  catalog,
 				Format:         "jsonline",
 				NoColor:        true,
 				SingleThreaded: test.singleThreaded,
