+++ packages/rslint/src/eslint-plugin/plugin/plugin-loader.ts
@@ -38,7 +38,7 @@ import {
 
 /**
  * Extension-aware config loader for worker init. Mirrors the strategy
- * `packages/rslint/src/config/config-loader.ts::loadConfigFile` uses on the
+ * `packages/rslint/src/config/config-file-loader.ts::loadConfigFile` uses on the
  * main thread so a `.ts`/`.mts`/`.cts` config file loads identically in both
  * places.
  *
@@ -207,7 +207,7 @@ export async function loadPluginsFromConfigFile(
   // Per entry, the live community plugins come from the object-form
   // `plugins` map (`plugins: { uc: unicornPlugin }`), selected via the
   // shared `selectPluginSource`. This mirrors `normalizeConfig` in
-  // packages/rslint/src/config/config-loader.ts — both paths must extract plugins
+  // packages/rslint/src/config/config-file-loader.ts — both paths must extract plugins
   // from the same source, or the worker holds a plugin set that doesn't
   // match what the main thread believed was active. Array-form `plugins`
   // (the native-name whitelist) carries no live objects and yields no
